Split local
