+++ frictionless/plugins/local/__init__.py
@@ -1,0 +1,3 @@
+from .control import LocalControl
+from .loader import LocalLoader
+from .plugin import LocalPlugin
